Move Customer, Plan and DataSource outside of Import namespace
