+++ fixtures/vcr_cassettes/ChartMogul_DataSource/API_Interactions/correctly_raises_errors_on_404.yml
@@ -2,7 +2,7 @@
 http_interactions:
 - request:
     method: delete
-    uri: https://api.chartmogul.com/v1/import/data_sources/1234-a-uuid-that-doesnt-exist
+    uri: https://api.chartmogul.com/v1/data_sources/1234-a-uuid-that-doesnt-exist
     body:
       encoding: US-ASCII
       string: ''
@@ -14,7 +14,7 @@ http_interactions:
   response:
     status:
       code: 404
-      message: 
+      message:
     headers:
       server:
       - nginx/1.9.10
@@ -41,6 +41,6 @@ http_interactions:
     body:
       encoding: UTF-8
       string: '{"error":"Data source not found"}'
-    http_version: 
+    http_version:
   recorded_at: Sun, 05 Jun 2016 17:05:01 GMT
 recorded_with: VCR 3.0.3
