+++ fixtures/vcr_cassettes/ChartMogul_DataSource/API_Interactions/retrieves_existing_data_source_matching_uuid.yml
@@ -1,0 +1,101 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.chartmogul.com/v1/data_sources
+    body:
+      encoding: UTF-8
+      string: '{"name":"TestDS"}'
+    headers:
+      User-Agent:
+      - Faraday v0.10.0
+      Content-Type:
+      - application/json
+      Authorization:
+      - Basic ZHVtbXktdG9rZW46ZHVtbXktdG9rZW4=
+  response:
+    status:
+      code: 201
+      message:
+    headers:
+      server:
+      - nginx/1.9.10
+      date:
+      - Sun, 05 Jun 2016 18:52:01 GMT
+      content-type:
+      - application/json; charset=utf-8
+      transfer-encoding:
+      - chunked
+      connection:
+      - close
+      x-frame-options:
+      - SAMEORIGIN
+      x-xss-protection:
+      - 1; mode=block
+      x-content-type-options:
+      - nosniff
+      etag:
+      - W/"5621b5bd639d8cf3876eddd3c3617ee5"
+      cache-control:
+      - max-age=0, private, must-revalidate
+      x-request-id:
+      - 67b3f0f2-9b9a-4bae-b631-578db1d7158b
+      x-runtime:
+      - '0.435113'
+      strict-transport-security:
+      - max-age=15768000
+    body:
+      encoding: UTF-8
+      string: '{"uuid":"ds_5ee8bf93-b0b4-4722-8a17-6b624a3af072","name":"TestDS","created_at":"2016-06-05T18:52:00.931Z","status":"never_imported"}'
+    http_version:
+  recorded_at: Tue, 03 Jan 2017 22:49:40 GMT
+- request:
+    method: get
+    uri: https://api.chartmogul.com/v1/data_sources/ds_5ee8bf93-b0b4-4722-8a17-6b624a3af072
+    body:
+      encoding: UTF-8
+      string: ''
+    headers:
+      User-Agent:
+      - Faraday v0.10.0
+      Content-Type:
+      - application/json
+      Authorization:
+      - Basic ZHVtbXktdG9rZW46ZHVtbXktdG9rZW4=
+  response:
+    status:
+      code: 200
+      message:
+    headers:
+      server:
+      - nginx/1.9.10
+      date:
+      - Sun, 05 Jun 2016 18:52:01 GMT
+      content-type:
+      - application/json; charset=utf-8
+      transfer-encoding:
+      - chunked
+      connection:
+      - close
+      x-frame-options:
+      - SAMEORIGIN
+      x-xss-protection:
+      - 1; mode=block
+      x-content-type-options:
+      - nosniff
+      etag:
+      - W/"5621b5bd639d8cf3876eddd3c3617ee5"
+      cache-control:
+      - max-age=0, private, must-revalidate
+      x-request-id:
+      - 67b3f0f2-9b9a-4bae-b631-578db1d7158b
+      x-runtime:
+      - '0.435113'
+      strict-transport-security:
+      - max-age=15768000
+    body:
+      encoding: UTF-8
+      string: '{"uuid":"ds_5ee8bf93-b0b4-4722-8a17-6b624a3af072","name":"TestDS","system": "Import API", "created_at":"2016-06-05T18:52:00.931Z","status":"never_imported"}'
+    http_version:
+  recorded_at: Tue, 03 Jan 2017 22:49:40 GMT
+recorded_with: VCR 3.0.3
