+++ fixtures/vcr_cassettes/ChartMogul_Plan/API_Interractions/correctly_handles_a_422_error.yml
@@ -2,17 +2,17 @@
 http_interactions:
 - request:
     method: post
-    uri: https://api.chartmogul.com/v1/data_sources
+    uri: https://api.chartmogul.com/v1/plans
     body:
       encoding: UTF-8
-      string: '{"name":null}'
+      string: '{"data_source_uuid":null,"name":null,"interval_count":null,"interval_unit":null,"external_id":null}'
     headers:
       User-Agent:
       - Faraday v0.9.1
       Content-Type:
       - application/json
       Authorization:
-      - Basic ZTk5MGJkMjNmNTQzYWNjZGY3ZDg4Yjk2OWNiMzAwOWI6NWM4YThlY2UxMTFmYmE4NTE1MGVmNzBjYmZmM2I3ZWI=
+      - Basic ODE2MjZkMWFiMzU5NTIwMzY5MjhjNjJjYmJiNDMyN2M6ODY3MDRmMDdkYzQwZmYzYjYyYzRjY2E5OTYxMjk2MzY=
   response:
     status:
       code: 422
@@ -21,7 +21,7 @@ http_interactions:
       server:
       - nginx/1.9.10
       date:
-      - Sun, 05 Jun 2016 17:04:50 GMT
+      - Sun, 05 Jun 2016 19:47:08 GMT
       content-type:
       - application/json; charset=utf-8
       transfer-encoding:
@@ -37,12 +37,15 @@ http_interactions:
       cache-control:
       - no-cache
       x-request-id:
-      - 76a25073-4c20-49fb-83b2-bbad2c361b0d
+      - 72fc5d2e-0e09-4d62-b0b6-c10299d79c0a
       x-runtime:
-      - '0.009928'
+      - '0.026567'
     body:
       encoding: UTF-8
-      string: '{"errors":{"name":"A name is required for each data source."}}'
+      string: '{"errors":{"name":"A unique name is required for each plan.","interval_count":"is
+        not a number","interval_unit":"Only ''day'', ''month'' or ''year'' are valid
+        values for interval_unit.","data_source_uuid":"A valid data source UUID is
+        required for each plan."}}'
     http_version:
-  recorded_at: Sun, 05 Jun 2016 17:04:50 GMT
+  recorded_at: Sun, 05 Jun 2016 19:47:08 GMT
 recorded_with: VCR 3.0.3
